Adapt Ymael to Windows
Ymael is now cross-platform GNU/Linux / Windows.
Modifications that were done:

  - modified secrets submodule to set the correct keyring
  - modified the notifier module to import and use the correct library
  - added ico file as the windows notifier does not handle png
  - made conditional imports for some libraries specific to linux of
  windows (using try/except ImportError)
  - made Core.export escape some characters in filename that will lead
  to issues in windows
  - specified the use of utf-8 where necessary
  - abandon the use of tmp files in Panexport due to PermissionError in
  windows
  - locked use of cli with a frozen exe in windows, as double click
  would activate the watch and not the gui (cli is still usable with
  sources)
  - updated main.spec with conditional var definitions for linux or
  windows
  - updated documentation
  - updated requirements and adapted makefile to the now conditional
  requirements files.

+++ ymael/core/export/panexport.py
@@ -2,7 +2,6 @@
 
 import pypandoc
 import os
-import tempfile
 import shutil
 import logging
 logger = logging.getLogger(__name__)
@@ -37,14 +36,17 @@ fontfamilyoptions: sfdefault
 ---
 
 """
-        tmp_file = tempfile.mkstemp(text=True)[1]
+        tmp_file = filename
+        if ext != ".md":
+            tmp_file += ".md"
         MDMaker(tmp_file, rps, add_before)
-        try:
-            pypandoc.convert_file(tmp_file, ext.lstrip("."), format="markdown",
-                    outputfile=filename,
-                    extra_args=['--pdf-engine', 'xelatex'])
-            os.remove(tmp_file)
-        except AttributeError:
-            shutil.move(tmp_file, filename+".md")
-            logger.exception("Conversion error. File is located at {}".format(filename+".md"))
-            raise
+        if ext != ".md":
+            try:
+                pypandoc.convert_file(tmp_file, ext.lstrip("."), format="markdown",
+                        outputfile=filename,
+                        extra_args=['--pdf-engine', 'xelatex'])
+                os.remove(tmp_file)
+            except AttributeError:
+                shutil.move(tmp_file, filename+".md")
+                logger.exception("Conversion error. File is located at {}".format(filename+".md"))
+                raise
